requirements.txt
pip>=21.0,<23.2
homeassistant==2023.3.0
colorlog==6.7.0
pre-commit==3.3.1
black==23.3.0
ruff==0.0.270
# These are duplicated in .pre-commit-config.yaml
reorder-python-imports==3.9.0
mypy==1.3.0
homeassistant-stubs==2023.3.0
types-python-slugify==8.0.0.2
voluptuous-stubs==0.1.1
# For mypy. Keep in sync with manifest.json. If changed, make sure subclasses in modbus_client are still valid!
pymodbus==3.2.2
